01-cursor-generated-chatbot/requirements.txt
streamlit==1.31.0
openai==1.3.0
python-dotenv==1.0.0 